+++ src/main/java/com/billing/yota/dao/AccountDao.java
@@ -2,11 +2,13 @@ package com.billing.yota.dao;
 
 import com.billing.yota.model.entity.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountDao {
 
     void create( Account account);
-    void update(long number, boolean isCan);
-    boolean checkTransfer(long number);
-    Account findByNumber(long number);
-    void changeBalance(long number, double balance);
+    void update(int number, boolean isCan);
+    boolean checkTransfer(int number);
+    Account findByNumber(int number);
+    void changeBalance(int number, BigDecimal balance);
 }
